+++ R/package.R
@@ -25,15 +25,15 @@ spark_dependencies <- function(spark_version, scala_version, ...) {
       latest <- read.table("http://s3.amazonaws.com/h2o-release/sparkling-water/rel-2.2/latest")
       sw_version <- sprintf("2.2.%s",latest)
       message(sprintf("Spark version %s detected. Will call latest Sparkling Water version %s",spark_version,sw_version))
-      if (packageVersion("h2o") != "3.16.0.2") {
-        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.2, which is compliant with the latest Sparkling Water version for Spark 2.2.* -> Sparkling Water version ', sw_version,'\n
-                       To update your h2o R package, copy/paste the following commands and then restart your R session:
+      if (packageVersion("h2o") != "3.16.0.4") {
+        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.4, which is compliant with the latest Sparkling Water version for Spark 2.2.* -> Sparkling Water version ', sw_version,'\n
+To update your h2o R package, copy/paste the following commands and then restart your R session:
                        
-                       detach("package:rsparkling", unload = TRUE)
-                       if ("package:h2o" %in% search()) { detach("package:h2o", unload = TRUE) }
-                       if (isNamespaceLoaded("h2o")){ unloadNamespace("h2o") }
-                       remove.packages("h2o")
-                       install.packages("h2o", type = "source", repos = "https://h2o-release.s3.amazonaws.com/h2o/rel-weierstrass/2/R")\n'))
+   detach("package:rsparkling", unload = TRUE)
+   if ("package:h2o" %in% search()) { detach("package:h2o", unload = TRUE) }
+   if (isNamespaceLoaded("h2o")){ unloadNamespace("h2o") }
+   remove.packages("h2o")
+   install.packages("h2o", type = "source", repos = "https://h2o-release.s3.amazonaws.com/h2o/rel-wheeler/4/R")\n'))
       }
     }
     else if (as.package_version(spark_version)$major == "2" && as.package_version(spark_version)$minor == "1") {
@@ -41,30 +41,30 @@ spark_dependencies <- function(spark_version, scala_version, ...) {
       latest <- read.table("http://s3.amazonaws.com/h2o-release/sparkling-water/rel-2.1/latest")
       sw_version <- sprintf("2.1.%s",latest)
       message(sprintf("Spark version %s detected. Will call latest Sparkling Water version %s",spark_version,sw_version))
-      if (packageVersion("h2o") != "3.16.0.2") {
-        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.2, which is compliant with the latest Sparkling Water version for Spark 2.1.* -> Sparkling Water version ', sw_version,'\n
+      if (packageVersion("h2o") != "3.16.0.4") {
+        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.4, which is compliant with the latest Sparkling Water version for Spark 2.1.* -> Sparkling Water version ', sw_version,'\n
 To update your h2o R package, copy/paste the following commands and then restart your R session:
                        
   detach("package:rsparkling", unload = TRUE)
   if ("package:h2o" %in% search()) { detach("package:h2o", unload = TRUE) }
   if (isNamespaceLoaded("h2o")){ unloadNamespace("h2o") }
   remove.packages("h2o")
-  install.packages("h2o", type = "source", repos =  "https://h2o-release.s3.amazonaws.com/h2o/rel-weierstrass/2/R")\n'))
+  install.packages("h2o", type = "source", repos =  "https://h2o-release.s3.amazonaws.com/h2o/rel-wheeler/4/R")\n'))
       }
     }else if (as.package_version(spark_version)$major == "2") {
       #Get latest Sparkling Water release for Spark 2.0.*
       latest <- read.table("http://s3.amazonaws.com/h2o-release/sparkling-water/rel-2.0/latest")
       sw_version <- sprintf("2.0.%s",latest)
       message(sprintf("Spark version %s detected. Will call latest Sparkling Water version %s",spark_version,sw_version))
-      if (packageVersion("h2o") != "3.16.0.2") {
-        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.2, which is compliant with the latest Sparkling Water version for Spark 2.0.* ->  Sparkling Water version ', sw_version,'\n
+      if (packageVersion("h2o") != "3.16.0.4") {
+        message(paste0('\nDetected H2O version ', packageVersion("h2o"),'. Please install H2O version 3.16.0.4, which is compliant with the latest Sparkling Water version for Spark 2.0.* ->  Sparkling Water version ', sw_version,'\n
 To update your h2o R package, copy/paste the following commands and then restart your R session:
 
   detach("package:rsparkling", unload = TRUE)
   if ("package:h2o" %in% search()) { detach("package:h2o", unload = TRUE) }
   if (isNamespaceLoaded("h2o")){ unloadNamespace("h2o") }
   remove.packages("h2o")
-  install.packages("h2o", type = "source", repos =  "https://h2o-release.s3.amazonaws.com/h2o/rel-weierstrass/2/R")\n'))
+  install.packages("h2o", type = "source", repos =  "https://h2o-release.s3.amazonaws.com/h2o/rel-wheeler/4/R")\n'))
       }
     } else if (as.package_version(spark_version)$major == "1" && as.package_version(spark_version)$minor == "6" ) { #Assuming Spark 1.6
       #Get latest Sparkling Water release for Spark 1.6.*